pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "art-interp"
version = "0.1.0"
description = "SAE-based interpretability tools for vision models - understand what AI sees in art"
authors = [
    {name = "Nathaniel Bahn", email = "nathaniel@example.com"},
]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
keywords = ["interpretability", "vision", "SAE", "CLIP", "machine learning", "art"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # Core ML
    "torch>=2.1.0",
    "open-clip-torch>=2.24.0",
    "huggingface_hub>=0.20.0",

    # Scientific computing
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",

    # Image processing
    "pillow>=10.0.0",
    "httpx>=0.26.0",

    # Progress and logging
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
server = [
    "fastapi>=0.109.0",
    "uvicorn>=0.25.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
]
labeling = [
    "google-generativeai>=0.3.0",
]
visualization = [
    "matplotlib>=3.8.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]
all = [
    "art-interp[server,labeling,visualization,dev]",
]

[project.urls]
Homepage = "https://github.com/yourusername/art-interp"
Documentation = "https://github.com/yourusername/art-interp#readme"
Repository = "https://github.com/yourusername/art-interp"

[project.scripts]
interp-extract = "interpretability.scripts.extract_features:main"
interp-monosemanticity = "interpretability.scripts.compute_monosemanticity:main"
interp-correlations = "interpretability.scripts.compute_correlations:main"
interp-server = "server.main:run"

[tool.hatch.build.targets.wheel]
packages = ["src/interpretability", "server"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
